Bypass confirm sign up when user is confirmed
