+++ packages/amplify-ui-components/src/components/amplify-s3-image-picker/amplify-s3-image-picker.tsx
@@ -1,7 +1,7 @@
 import { Component, Prop, h, State, Host } from '@stencil/core';
 import { Logger, I18n } from '@aws-amplify/core';
 import { AccessLevel } from '../../common/types/storage-types';
-import { calcKey, getStorageObject, putStorageObject } from '../../common/storage-helper';
+import { calcKey, getStorageObject, putStorageObject } from '../../common/storage-helpers';
 import { Translations } from '../../common/Translations';
 
 const logger = new Logger('S3ImagePicker');
